refactor(readr): move `SITE_TITLE` and `DEFAULT_POST_IMAGE_PATH` from config to constant
