+++ packages/readr/constants/constant.ts
@@ -1,0 +1,6 @@
+// 這裡管理固定數值
+
+const SITE_TITLE = 'READr 讀+'
+const DEFAULT_POST_IMAGE_PATH = '/icons/default/post.svg'
+
+export { DEFAULT_POST_IMAGE_PATH, SITE_TITLE }
